Fixed Inconsistent Doublet Generation
- Added sleep(1) to flight simulation thread when in analysis mode; this
seems to give the flight controls thread a chance to obtain the correct
value of time for generating doublets

- Added AtomicInteger to Integrate6DOFEquations, which handles time in
milliseconds outside of the flight simulation thread. Inside, double
values in seconds still handle time for integrators and thread loops

+++ src/com/chrisali/javaflightsim/simulation/interfaces/SimulationController.java
@@ -1,5 +1,7 @@
 package com.chrisali.javaflightsim.simulation.interfaces;
 
+import java.util.concurrent.atomic.AtomicInteger;
+
 import com.chrisali.javaflightsim.initializer.LWJGLSwingSimulationController;
 import com.chrisali.javaflightsim.simulation.integration.Integrate6DOFEquations;
 import com.chrisali.javaflightsim.simulation.setup.SimulationConfiguration;
@@ -19,7 +21,10 @@ public interface SimulationController {
 	
 	public Integrate6DOFEquations getSimulation();
 	
-	public double getTime();
+	/**
+	 * @return current elapsed time in the simulation in millisec
+	 */
+	public AtomicInteger getTime();
 	
 	public void plotSimulation();
